Migrate Babel plugins to Rollup plugins
