+++ packages/prerender/package.json
@@ -61,10 +61,13 @@
   },
   "devDependencies": {
     "@cedarjs/framework-tools": "workspace:*",
+    "@rollup/plugin-swc": "^0.4.0",
+    "@swc/core": "^1.12.1",
     "@types/mime-types": "2.1.4",
     "babel-plugin-tester": "11.0.4",
     "concurrently": "8.2.2",
     "publint": "0.3.12",
+    "rollup-plugin-swc3": "^0.12.1",
     "tsx": "4.19.4",
     "typescript": "5.6.2",
     "vitest": "2.1.9"
