+++ packages/prerender/src/runPrerender.tsx
@@ -3,18 +3,17 @@ import path from 'path'
 
 import React from 'react'
 
+// @ts-expect-error CJS vs ESM
 import { ApolloClient, InMemoryCache } from '@apollo/client'
-import alias from '@rollup/plugin-alias'
 import type { CheerioAPI } from 'cheerio'
 import { load as loadHtml } from 'cheerio'
 import ReactDOMServer from 'react-dom/server'
-import unimportPlugin from 'unimport/unplugin'
 
 import {
   registerApiSideBabelHook,
   registerWebSideBabelHook,
 } from '@cedarjs/babel-config'
-import { getPaths, ensurePosixPath, getConfig } from '@cedarjs/project-config'
+import { getPaths, ensurePosixPath } from '@cedarjs/project-config'
 import { LocationProvider } from '@cedarjs/router'
 import { matchPath } from '@cedarjs/router/dist/util'
 import type { QueryInfo } from '@cedarjs/web'
@@ -28,15 +27,8 @@ import {
   PrerenderGqlError,
 } from './errors'
 import { executeQuery, getGqlHandler } from './graphql/graphql'
-import {
-  getPathsFromTypeScriptConfig,
-  getRootHtmlPath,
-  parseTypeScriptConfigFiles,
-  registerShims,
-  writeToDist,
-} from './internal'
+import { getRootHtmlPath, registerShims, writeToDist } from './internal'
 import { rollupRequire } from './rollup-require/rollup-require'
-import { cedarjsDirectoryNamedImportPlugin } from './rollupPlugins/rollup-plugin-cedarjs-directory-named-imports'
 
 // Create an apollo client that we can use to prepopulate the cache and restore it client-side
 const prerenderApolloClient = new ApolloClient({ cache: new InMemoryCache() })
@@ -308,101 +300,26 @@ export const runPrerender = async ({
     },
   })
 
-  // Need the project config to know if trusted graphql documents is being used
-  // and decide to use the gql tag import or the trusted document gql function
-  // generated by code gen client preset
-  const config = getConfig()
-  const useTrustedDocumentsGqlTag = config.graphql.trustedDocuments
-
-  const tsConfigs = parseTypeScriptConfigFiles()
-  const webBase = getPaths().web.base
-
+  console.log('')
+  console.log()
+  console.log('before rollupRequire\n\n')
   const { mod: App } = await rollupRequire({
     cwd: getPaths().web.base,
     filepath: getPaths().web.app,
-    rollupOptions: {
-      plugins: [
-        alias({
-          entries: [
-            {
-              find: 'src',
-              replacement: getPaths().web.src,
-            },
-            {
-              find: '$api',
-              replacement: getPaths().api.base,
-            },
-            ...getPathsFromTypeScriptConfig(tsConfigs.web, webBase),
-          ],
-        }),
-        cedarjsDirectoryNamedImportPlugin(),
-        unimportPlugin.rollup({
-          imports: [
-            // import React from 'react'
-            {
-              name: 'default',
-              as: 'React',
-              from: 'react',
-            },
-            // import gql from 'graphql-tag'
-            !useTrustedDocumentsGqlTag && {
-              name: 'default',
-              as: 'gql',
-              from: 'graphql-tag',
-            },
-            // import { gql } from 'src/graphql/gql'
-            useTrustedDocumentsGqlTag && {
-              name: 'gql',
-              from: `web/src/graphql/gql`,
-            },
-          ].filter(<T,>(v?: T | false): v is T => Boolean(v)),
-        }),
-      ],
-    },
   })
 
+  console.log('\nApp\n', App({}))
+
   const { mod: Routes } = await rollupRequire({
+    cwd: getPaths().web.base,
     filepath: getPaths().web.routes,
-    rollupOptions: {
-      plugins: [
-        alias({
-          entries: [
-            {
-              find: 'src',
-              replacement: getPaths().web.src,
-            },
-            {
-              find: '$api',
-              replacement: getPaths().api.base,
-            },
-            ...getPathsFromTypeScriptConfig(tsConfigs.web, webBase),
-          ],
-        }),
-        cedarjsDirectoryNamedImportPlugin(),
-        unimportPlugin.rollup({
-          imports: [
-            // import React from 'react'
-            {
-              name: 'default',
-              as: 'React',
-              from: 'react',
-            },
-            // import gql from 'graphql-tag'
-            !useTrustedDocumentsGqlTag && {
-              name: 'default',
-              as: 'gql',
-              from: 'graphql-tag',
-            },
-            // import { gql } from 'src/graphql/gql'
-            useTrustedDocumentsGqlTag && {
-              name: 'gql',
-              from: `web/src/graphql/gql`,
-            },
-          ].filter(<T,>(v?: T | false): v is T => Boolean(v)),
-        }),
-      ],
-    },
   })
+
+  console.log('\nRoutes\n', Routes({}))
+
+  if (Math.random() < 5) {
+    throw new Error('Happy Exit')
+  }
 
   const componentAsHtml = await recursivelyRender(
     App.default,
